Improve project files structure
