+++ Zotero/Scenes/Detail/ReaderCoordinator.swift
@@ -10,25 +10,6 @@ import UIKit
 import CocoaLumberjackSwift
 import RxSwift
 
-protocol ReaderAnnotation {
-    var key: String { get }
-    var type: AnnotationType { get }
-    var pageLabel: String { get }
-    var lineWidth: CGFloat? { get }
-    var color: String { get }
-    var comment: String { get }
-    var text: String? { get }
-    var fontSize: CGFloat? { get }
-    var sortIndex: String { get }
-    var dateAdded: Date { get }
-    var dateModified: Date { get }
-    var tags: [Tag] { get}
-
-    func author(displayName: String, username: String) -> String
-    func isAuthor(currentUserId: Int) -> Bool
-    func editability(currentUserId: Int, library: Library) -> AnnotationEditability
-}
-
 protocol ReaderError: Error {
     var title: String { get }
     var message: String { get }
